fix memory leak
